+++ extensions/amazon-aws/src/components/codepipeline/toggle-stage-transition-action.tsx
@@ -3,21 +3,20 @@ import {
   CodePipelineClient,
   DisableStageTransitionCommand,
   EnableStageTransitionCommand,
-  StageState,
   StageTransitionType,
 } from "@aws-sdk/client-codepipeline";
 import { getErrorMessage } from "../../util";
+import { MutatePromise } from "@raycast/utils";
+import { Pipeline, PipelineStage } from "../../codepipeline";
 
 export const ToggleStageTransitionAction = ({
-  pipelineName,
+  pipeline,
   isLoading,
-  revalidatePipeline,
-  stages,
+  mutate,
 }: {
-  pipelineName: string;
+  pipeline: Pipeline;
   isLoading: boolean;
-  revalidatePipeline: () => void;
-  stages?: (StageState & { nextStage?: StageState })[];
+  mutate: MutatePromise<Pipeline[]>;
 }) => {
   return (
     <ActionPanel.Submenu
@@ -27,20 +26,20 @@ export const ToggleStageTransitionAction = ({
       isLoading={isLoading}
       filtering
     >
-      {(stages ?? [])
+      {pipeline.stages
         .filter((s) => !!s.nextStage)
         .map((s) => {
           const transitionEnabled = !!s.nextStage!.inboundTransitionState?.enabled;
           return (
             <Action
-              key={`${pipelineName}-${s.stageName}-${s.nextStage!.stageName}`}
+              key={`${pipeline.name}-${s.stageName}-${s.nextStage!.stageName}`}
               icon={
                 transitionEnabled
                   ? { source: Icon.Bolt, tintColor: Color.Green }
                   : { source: Icon.BoltDisabled, tintColor: Color.Red }
               }
               title={`${s.stageName} -> ${s.nextStage!.stageName}`}
-              onAction={async () => toggleStageTransition(transitionEnabled, pipelineName, s, revalidatePipeline)}
+              onAction={() => toggleStageTransition(transitionEnabled, pipeline.name!, s, mutate)}
             />
           );
         })}
@@ -51,59 +50,86 @@ export const ToggleStageTransitionAction = ({
 const toggleStageTransition = async (
   transitionEnabled: boolean,
   pipelineName: string,
-  s: { stageName?: string; nextStage?: { stageName?: string } },
-  revalidatePipeline: () => void,
+  stage: PipelineStage,
+  mutate: MutatePromise<Pipeline[]>,
 ) => {
-  if (transitionEnabled) {
-    const toast = await showToast(
-      Toast.Style.Animated,
-      "❗Disabling transition",
-      `between ${s.stageName} -> ${s.nextStage!.stageName}`,
-    );
-    new CodePipelineClient({})
-      .send(
-        new DisableStageTransitionCommand({
-          pipelineName,
-          stageName: s.nextStage!.stageName,
-          transitionType: StageTransitionType.Inbound,
-          reason: "Disabled by Raycast",
-        }),
-      )
-      .then(() => {
-        toast.style = Toast.Style.Success;
-        toast.title = "✅ Disabled transition";
-      })
-      .catch((err) => {
-        captureException(err);
-        toast.style = Toast.Style.Failure;
-        toast.title = "❌ Failed to disable transition";
-        toast.message = getErrorMessage(err);
-      })
-      .finally(revalidatePipeline);
-  } else {
-    const toast = await showToast(
-      Toast.Style.Animated,
-      "❗Enabling transition",
-      `between ${s.stageName} -> ${s.nextStage!.stageName}`,
-    );
-    new CodePipelineClient({})
-      .send(
-        new EnableStageTransitionCommand({
-          pipelineName,
-          stageName: s.nextStage!.stageName,
-          transitionType: StageTransitionType.Inbound,
-        }),
-      )
-      .then(() => {
-        toast.style = Toast.Style.Success;
-        toast.title = "✅ Enabled transition";
-      })
-      .catch((err) => {
-        captureException(err);
-        toast.style = Toast.Style.Failure;
-        toast.title = "❌ Failed to enable transition";
-        toast.message = getErrorMessage(err);
-      })
-      .finally(revalidatePipeline);
-  }
+  await mutate(transitionEnabled ? disableTransition(pipelineName, stage) : enableTransition(pipelineName, stage), {
+    optimisticUpdate: (pipelines) =>
+      pipelines.map((p) =>
+        p.name !== pipelineName
+          ? p
+          : {
+              ...p,
+              stages: p.stages.map((s) =>
+                s.stageName !== stage.stageName
+                  ? s
+                  : {
+                      ...s,
+                      nextStage: {
+                        ...s.nextStage,
+                        inboundTransitionState: { ...s.nextStage?.inboundTransitionState, enabled: !transitionEnabled },
+                      },
+                    },
+              ),
+            },
+      ),
+    shouldRevalidateAfter: false,
+  });
+};
+
+const enableTransition = async (pipelineName: string, stage: PipelineStage) => {
+  const toast = await showToast(
+    Toast.Style.Animated,
+    "❗Enabling transition",
+    `between ${stage.stageName} -> ${stage.nextStage!.stageName}`,
+  );
+  return await new CodePipelineClient({})
+    .send(
+      new EnableStageTransitionCommand({
+        pipelineName,
+        stageName: stage.nextStage!.stageName,
+        transitionType: StageTransitionType.Inbound,
+      }),
+    )
+    .then(() => {
+      toast.style = Toast.Style.Success;
+      toast.title = "✅ Enabled transition";
+      return true;
+    })
+    .catch((err) => {
+      captureException(err);
+      toast.style = Toast.Style.Failure;
+      toast.title = "❌ Failed to enable transition";
+      toast.message = getErrorMessage(err);
+      throw err;
+    });
+};
+
+const disableTransition = async (pipelineName: string, stage: PipelineStage) => {
+  const toast = await showToast(
+    Toast.Style.Animated,
+    "❗Disabling transition",
+    `between ${stage.stageName} -> ${stage.nextStage!.stageName}`,
+  );
+  return await new CodePipelineClient({})
+    .send(
+      new DisableStageTransitionCommand({
+        pipelineName,
+        stageName: stage.nextStage!.stageName,
+        transitionType: StageTransitionType.Inbound,
+        reason: "Disabled by Raycast",
+      }),
+    )
+    .then(() => {
+      toast.style = Toast.Style.Success;
+      toast.title = "✅ Disabled transition";
+      return true;
+    })
+    .catch((err) => {
+      captureException(err);
+      toast.style = Toast.Style.Failure;
+      toast.title = "❌ Failed to disable transition";
+      toast.message = getErrorMessage(err);
+      throw err;
+    });
 };