+++ extensions/open-link-in-specific-browser/src/utils/input-utils.ts
@@ -1,7 +1,9 @@
-import { getSelectedText, Clipboard } from "@raycast/api";
+import { getSelectedText, Clipboard, getFrontmostApplication } from "@raycast/api";
 import { isEmpty, isUrl, urlIPBuilder } from "./common-utils";
 import { isIP } from "net";
 import { ItemSource, ItemType } from "../types/types";
+import { chromiumBrowserNames, webkitBrowserNames } from "./constants";
+import { getChromiumBrowserPath, getWebkitBrowserPath } from "./applescript-utils";
 
 export class ItemInput {
   content: string;
@@ -58,7 +60,7 @@ const clipboard = async () => {
 
 export const fetchItemInput = async () => {
   const itemInput = new ItemInput();
-  return await getSelectedText()
+  const inputText = await getSelectedText()
     .then(async (text) =>
       !isEmpty(text)
         ? itemInput.setContent(text.substring(0, 9999)).setSource(ItemSource.SELECTED).setType()
@@ -77,4 +79,24 @@ export const fetchItemInput = async () => {
       !isEmpty(item.content) ? itemInput : itemInput.setContent("").setSource(ItemSource.NULL).setType(),
     )
     .catch(() => itemInput.setContent("").setSource(ItemSource.NULL).setType());
+
+  if (inputText.type !== ItemType.URL) {
+    const frontmostAppUrl = await getFrontmostAppUrl();
+    if (!isEmpty(frontmostAppUrl)) {
+      return new ItemInput(frontmostAppUrl, ItemSource.FRONTMOSTAPP, ItemType.URL);
+    }
+  }
+  return inputText;
+};
+
+export const getFrontmostAppUrl = async () => {
+  const frontmostApp = await getFrontmostApplication();
+  // get browser web page url
+  let url = "";
+  if (webkitBrowserNames.includes(frontmostApp.name)) {
+    url = await getWebkitBrowserPath(frontmostApp.name);
+  } else if (chromiumBrowserNames.includes(frontmostApp.name)) {
+    url = await getChromiumBrowserPath(frontmostApp.name);
+  }
+  return url;
 };