+++ extensions/paste-as-plain-text/src/paste-as-plain-text.tsx
@@ -1,5 +1,5 @@
-import { captureException, LaunchProps, showHUD } from "@raycast/api";
-import { getArgument, isEmpty } from "./utils/common-utils";
+import { captureException, LaunchProps, Toast } from "@raycast/api";
+import { getArgument, isEmpty, showCustomHUD } from "./utils/common-utils";
 import { PasteFormat } from "./types/types";
 import { pasteAs } from "./utils/paste-as-utils";
 
@@ -15,6 +15,6 @@ export default async (props: LaunchProps<{ arguments: PasteAsArguments }>) => {
   } catch (e) {
     captureException(e);
     console.error(e);
-    await showHUD("🚨 " + e);
+    await showCustomHUD({ title: String(e), style: Toast.Style.Failure });
   }
 };