+++ extensions/pins/src/lib/preferences.ts
@@ -34,6 +34,11 @@ export interface ExtensionPreferences {
    * The default sort strategy for lists of pins outside of groups.
    */
   defaultSortStrategy: string;
+
+  /**
+   * The default Raycast AI model to use for AI queries.
+   */
+  defaultAIModel: string;
 }
 
 /**
@@ -194,6 +199,11 @@ export interface ViewPinsPreferences {
    * Whether to display the tags of each pin as accessories.
    */
   showTags: boolean;
+
+  /**
+   * Whether to display the number of linked pins as an accessory.
+   */
+  showLinkCount: boolean;
 }
 
 /**