+++ extensions/yoink/src/constants.ts
@@ -1,0 +1,3 @@
+export const YoinkBundleIdentifier = "at.EternalStorms.Yoink";
+export const YoinkPath = "/Applications/Yoink.app";
+export const YoinkAppStoreLink = "https://apps.apple.com/cn/app/yoink-better-drag-and-drop/id457622435?l=en-GB&mt=12";