+++ extensions/amazon-aws/CHANGELOG.md
@@ -1,10 +1,13 @@
 # Amazon AWS Changelog
 
+## [Improvements] - {PR_MERGE_DATE}
 
-## [GlueJobs command] - {PR_MERGE_DATE}
-
-- Added Glue Jobs command to list Glue Jobs
 - Changed AwsAction.Console component, to create sso shortcut login links, if possible
+
+## [Fixes] - 2024-06-24
+
+- Fixed [the issue](https://github.com/raycast/extensions/issues/13088) when codepipeline(s) are rendered with no executions.
+- Adds mutate optimization for toggling stage transition.
 
 ## [Console command] - 2024-06-19