+++ extensions/ugly-face/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Ugly Face Changelog
 
+## [Feat] - 2024-06-27
+
+- Added a copy action.
+
 ## [Feat] - 2024-06-04
 - Added custom configuration.